Moving google map up one level to not use a middle man component
